requirements.txt
torch>=1.9.0
torchaudio>=0.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
pyyaml>=5.4.0

# Evaluation Metrics
pesq>=0.0.3
pystoi>=0.3.3
mbstoi @ git+https://github.com/HolyWu/mbstoi.git  # if using MBSTOI module

# Config & CLI
hydra-core>=1.1.0
omegaconf>=2.1.0

# Audio I/O
soundfile>=0.10.3.post1  # Provides the SoundFile class (pysoundfile)

# Utilities
tqdm>=4.60.0
scipy>=1.6.0
numpy>=1.19.0
